fix wrong glob in #fetch_movefile
Thanks to a failure on the CI system I discovered
that we was doing a glob file search for the Movefile*
in order to find in order to honorate Movefile.yaml
also.

This behaviour was leading to confusion when e.g. a
Movefile_old was present. Moreover the globstar was
returning multiple files in a system dependent order,
and so when we was taking the `.first` of the match
we had unpredictable results.

New glob pattern `Movefile{,.yml}` should do the trick.