Refactor Deployer::Base and Movefile classes. See long desc.
Deployer::Base used to be a god class, bringing too much logic
and with an overloaded .deployer_for method. We already moved
`fetch` logic from there to the new Movefile class.

Now we have 3 big changes:

1. Deployer::Base can no more guess the environment based on
movefile.yml and optional `--environment` cli argument. We moved
the logic inside Movefile.environment instance method. It optionally
want the cli_options object, in order to understand if the user required
a specific environment to be used. If not, as per the previous logic, it
returns the only environment configured or a error if we have more than
one, requesting the user to pass `--environment` arg to the cli.

2. Movefile.fetch instance method now returns an object with deep
symbolized keys. This brought changes to both code and tests, but also
a unified way to access all the options (those fetched from the movefile
and those populated by Thor). I think moreover using symbols is a less
error prone and solid approach

3. Now custom movefile name and start_dir (for traversing) are no more
arguments of Movefile.fetch method, but of Movefile class initializer.
This way we save them in instance variables and made them available to both
.fetch and .environment methods. Changes to code and tests required and
done. I've noticed that this way I removed some mocking from specs and I
am happy with it

3.a Movefile.fetch lost its arguments, but acquaired a new boolean arg, to
switch off it's verbosity. Now that we fetch the movefile.yml within many
different classes, sometime it's not desired to print to STDOUT info to the
user.