should fix #428
Here we completely subverted the way we produce rsync command.
Managing subdirectories transfer is now done with rsync's includes
directives and the rsync tranfer root is always the wordpress root.

This way an exclusion such as `wp-content/plugins/myplugin` in the
movefile.yml is always understood from wormdove/rsync when push/pull-ing
`--all`, core, or even a single component.

ref: https://github.com/welaika/wordmove/issues/428