Hooks raise when their command exits with error
This is a new default behaviour, so it is considered
not to be backward compatible, because movefile.yml
that previously worked, now could bring Wordmove to
drop exceptions, thus breaking deployment scenarios.

This incompatibility will **only affect users who
was knowingly having failing hooks in their hooks
chain**.

We have thought about this change a lot and, even if
opinionated, we think it will bring some sanity and
will prevent high risk broken deploys.
Actually a hook chain is commonly considered to have
to be an all "green" chain. If you fail pre-push hooks
you do not want to deploy the code (linter failed?
webpack compilation failed? the code is broken).
And if you have a failed post-push hook you can at
least grab Wordmove's non-zero exit status and do
some other action based on that.

The base concept is that if you extend the deploy
process with hooks, that means you need more commands
to orchestrate and complete it. If one step fail,
the deploy is failed - said that post-push hooks
cannot stop the actual Wordmove's code transfer.

This new behaviour also opens the door to new features
(such as an integrated notification system) and to
better integration in a more
continuous-delivery-oriented approach.