Refactor SqlAdapter::Wpcli spec
Generally what I'm trying to do is to
avoid "overmocking". There were too much
stub we was relying on, but I opted to
maintain only the one about invoking the
`wp` command, so that we don't need to have
the binary in the repo for running tests.

I've also corercted the texts and reorganized
the contexts with a different nesting. Now they
look like this:

```
Wordmove::SqlAdapter::Wpcli
  #command
     having wp-cli.yml in local_path
      returns the right command as a string
    without wp-cli.yml in local_path
      but still reachable by wp-cli
        returns the right command as a string
    without any wp-cli configuration
      returns the right command with '--path' flag set to local_path
```